Prevent GKE cleanup from erroring
During the "finish" cleanup of the GKE tests, we delete images.  When
running tests in parallel (I believe this is the cause), you may see an
error like:

    ERROR: (gcloud.container.images.delete) Not found: response:

if the script attempts to delete an already-deleted image.  This is not
an important error, and we don't want our test to fail because of it.
So we turn of "set -e" when doing these deletions.